correction
git-svn-id: svn+ssh://svn.cern.ch/reps/dirac/LHCbDIRAC/trunk@9855 4525493e-7705-40b1-a816-d608a930855b

+++ LHCbDIRAC/BookkeepingSystem/Client/BookkeepingClient.py
@@ -139,7 +139,7 @@ class BookkeepingClient:
   
   #############################################################################
   def getProductionsWithEventTypes(self, eventType, configName,  configVersion, processingPass):
-    return self.server.getProductionsWithEventTypes(int(eventType), configName,  configVersion, processingPass)
+    return self.server.getProductionsWithEventTypes(long(eventType), configName,  configVersion, processingPass)
   
   #############################################################################
